authsite/requirements.txt
Django
django-cors-headers
uWSGI
certbot
certbot-nginx
cmake
face_recognition
Pillow
pylibmc
opencv-python
